requirements.txt
pandas
pytest
playwright
requests
psycopg2-binary
pyyaml